global: fixes oauth issues
