+++ requirements.txt
@@ -1,3 +1,3 @@
--e git+git://github.com/inveniosoftware/invenio-oauthclient.git@14bbc6c136f1466aafda4891313c563503cab95e#egg=invenio-oauthclient
+-e git+git://github.com/ioannistsanaktsidis/invenio-oauthclient.git@f6f438239f356a994ae6bbf0f399c5c0edbdc008#egg=invenio-oauthclient
 
 -e .[all]
